move parser class to lib/parser
